+++ backend/controllers/recipesController.js
@@ -1,64 +1,65 @@
-const Recipe = require('../models/recipesData.js');
-const axios = require('axios');
+const Recipe = require("../models/recipesData.js");
+const axios = require("axios");
+const transformDataForDB = require("../DB_insert_algorithm/transformDataForDB.js");
 
-const transformDataForDB = (data) => {
-    const transformedMeals = [];
-
-    for (let i = 0; i < data.meals.length; i++) {
-        const meal = data.meals[i];
-        const ingredients = [];
-
-        for (let j = 1; j <= 20; j++) {
-            const ingredient = meal[`strIngredient${j}`];
-            const measure = meal[`strMeasure${j}`];
-
-            if (ingredient && ingredient.trim() !== "") {
-                ingredients.push(measure + ' ' + ingredient);
-            }
-        }
-
-        const instructions = [];
-        const rawInstructions = meal.strInstructions.split('.');
-        for (let instruction of rawInstructions) {
-            const trimmedInstruction = instruction.trim();
-            if (trimmedInstruction) {
-                instructions.push(trimmedInstruction);
-            }
-        }
-
-        const transformedMeal = {
-            name: meal.strMeal,
-            catagory: meal.strCategory,
-            cuisine: meal.strArea,
-            picutre: meal.strMealThumb,
-            ingredients: ingredients,
-            instructions:instructions,
-            tags: meal.strTags ?meal.strTags.split(','):[],
-            youtubeLink: meal.strYoutube
-        };
-
-        transformedMeals.push(transformedMeal);
-    }
-
-    return transformedMeals;
-};
 exports.insertRecipes = async (req, res) => {
-    try {
-        const letters = 'abcdefghijklmnopqrstuvwxyz'.split('');
-        const allMeals = [];
+  try {
+    const letters = "abcdefghijklmnopqrstuvwxyz".split("");
+    const allMeals = [];
 
-        for (let letter of letters) {
-            const response = await axios.get(`https://www.themealdb.com/api/json/v1/1/search.php?f=${letter}`);
-            if (response.data.meals) {
-                allMeals.push(...response.data.meals);
-            }
-        }
-
-        const transformedData = transformDataForDB({ meals: allMeals });
-
-        //await Recipe.insertMany(transformedData);
-        res.status(200).json({ message: 'Recipes inserted successfully' });
-    } catch (error) {
-        res.status(500).json({ message: 'Error inserting recipes', error });
+    for (let letter of letters) {
+      const response = await axios.get(
+        `https://www.themealdb.com/api/json/v1/1/search.php?f=${letter}`
+      );
+      if (response.data.meals) {
+        allMeals.push(...response.data.meals);
+      }
     }
+
+    const transformedData = transformDataForDB({ meals: allMeals });
+
+    //await Recipe.insertMany(transformedData);
+    res.status(200).json({ message: "Recipes inserted successfully" });
+  } catch (error) {
+    return next({
+      message: "error in insertRecipes: " + err,
+      log: err,
+    });
+  }
+};
+
+exports.getRecipeByName = async (req, res) => {
+  const { name } = req.body;
+console.log("hit")
+  try {
+    const recipe = await Recipe.findOne({ name: new RegExp(name, "i") });
+    if (recipe) {
+      res.status(200).json(recipe);
+    } else {
+      res.status(404).json({ message: "Recipe not found" });
+    }
+  } catch (error) {
+    return next({
+      message: "error in getRecipeByName: " + err,
+      log: err,
+    });
+  }
+};
+
+exports.getRecipeById = async (req, res) => {
+    const { id } = req.params;
+
+    try {
+        const recipe = await Recipe.findById(id);
+        if (recipe) {
+            res.status(200).json(recipe);
+        } else {
+            res.status(404).json({ message: 'Recipe not found' });
+        }
+    }  catch (error) {
+        return next({
+          message: "error in getRecipeById: " + err,
+          log: err,
+        });
+      }
 };
